Expand graphics id to u16 to allow more than 255 graphics objects.

+++ src/union_room_player_avatar.c
@@ -19,25 +19,25 @@ static u8 CreateTask_AnimateUnionRoomPlayers(void);
 static u32 IsUnionRoomPlayerInvisible(u32, u32);
 static void SetUnionRoomObjectFacingDirection(s32, s32, u8);
 
-static const u8 sUnionRoomObjGfxIds[GENDER_COUNT][MAX_UNION_ROOM_PLAYERS + 2] = {
+static const u16 sUnionRoomObjGfxIds[GENDER_COUNT][MAX_UNION_ROOM_PLAYERS + 2] = {
     [MALE] = {
-        OBJ_EVENT_GFX_MAN_3, 
-        OBJ_EVENT_GFX_BLACK_BELT, 
-        OBJ_EVENT_GFX_CAMPER, 
-        OBJ_EVENT_GFX_YOUNGSTER, 
-        OBJ_EVENT_GFX_PSYCHIC_M, 
-        OBJ_EVENT_GFX_BUG_CATCHER, 
-        OBJ_EVENT_GFX_MAN_4, 
+        OBJ_EVENT_GFX_MAN_3,
+        OBJ_EVENT_GFX_BLACK_BELT,
+        OBJ_EVENT_GFX_CAMPER,
+        OBJ_EVENT_GFX_YOUNGSTER,
+        OBJ_EVENT_GFX_PSYCHIC_M,
+        OBJ_EVENT_GFX_BUG_CATCHER,
+        OBJ_EVENT_GFX_MAN_4,
         OBJ_EVENT_GFX_MAN_5
     },
     [FEMALE] = {
-        OBJ_EVENT_GFX_WOMAN_5, 
-        OBJ_EVENT_GFX_HEX_MANIAC, 
-        OBJ_EVENT_GFX_PICNICKER, 
-        OBJ_EVENT_GFX_LASS, 
-        OBJ_EVENT_GFX_LASS, 
-        OBJ_EVENT_GFX_GIRL_3, 
-        OBJ_EVENT_GFX_WOMAN_2, 
+        OBJ_EVENT_GFX_WOMAN_5,
+        OBJ_EVENT_GFX_HEX_MANIAC,
+        OBJ_EVENT_GFX_PICNICKER,
+        OBJ_EVENT_GFX_LASS,
+        OBJ_EVENT_GFX_LASS,
+        OBJ_EVENT_GFX_GIRL_3,
+        OBJ_EVENT_GFX_WOMAN_2,
         OBJ_EVENT_GFX_BEAUTY
     }
 };
@@ -90,13 +90,13 @@ static const u8 sUnionRoomLocalIds[] = {
 
 // Unused
 static const u16 sHidePlayerFlags[] = {
-    FLAG_HIDE_UNION_ROOM_PLAYER_1, 
-    FLAG_HIDE_UNION_ROOM_PLAYER_2, 
-    FLAG_HIDE_UNION_ROOM_PLAYER_3, 
-    FLAG_HIDE_UNION_ROOM_PLAYER_4, 
-    FLAG_HIDE_UNION_ROOM_PLAYER_5, 
-    FLAG_HIDE_UNION_ROOM_PLAYER_6, 
-    FLAG_HIDE_UNION_ROOM_PLAYER_7, 
+    FLAG_HIDE_UNION_ROOM_PLAYER_1,
+    FLAG_HIDE_UNION_ROOM_PLAYER_2,
+    FLAG_HIDE_UNION_ROOM_PLAYER_3,
+    FLAG_HIDE_UNION_ROOM_PLAYER_4,
+    FLAG_HIDE_UNION_ROOM_PLAYER_5,
+    FLAG_HIDE_UNION_ROOM_PLAYER_6,
+    FLAG_HIDE_UNION_ROOM_PLAYER_7,
     FLAG_HIDE_UNION_ROOM_PLAYER_8
 };
 
@@ -118,7 +118,7 @@ static bool32 IsPlayerStandingStill(void)
         return FALSE;
 }
 
-static u8 GetUnionRoomPlayerGraphicsId(u32 gender, u32 id)
+static u16 GetUnionRoomPlayerGraphicsId(u32 gender, u32 id)
 {
     return sUnionRoomObjGfxIds[gender][id % MAX_UNION_ROOM_PLAYERS];
 }
@@ -413,10 +413,10 @@ void CreateGroupMemberSpritesInvisible(u8 * spriteIds, s32 playerIdx)
     for (direction = DIR_NONE; direction <= DIR_EAST; direction++)
     {
         s32 id = UR_PLAYER_SPRITE_ID(playerIdx, direction);
-        spriteIds[id] = CreateObjectSprite(OBJ_EVENT_GFX_MAN_4, 
-                                           id - UR_SPRITE_START_ID, 
-                                           sUnionRoomPlayerCoords[playerIdx][0] + sFacingDirectionOffsets[direction][0], 
-                                           sUnionRoomPlayerCoords[playerIdx][1] + sFacingDirectionOffsets[direction][1], 
+        spriteIds[id] = CreateObjectSprite(OBJ_EVENT_GFX_MAN_4,
+                                           id - UR_SPRITE_START_ID,
+                                           sUnionRoomPlayerCoords[playerIdx][0] + sFacingDirectionOffsets[direction][0],
+                                           sUnionRoomPlayerCoords[playerIdx][1] + sFacingDirectionOffsets[direction][1],
                                            3, 1);
         SetObjectEventSpriteInvisibility(id - UR_SPRITE_START_ID, TRUE);
     }
@@ -457,7 +457,7 @@ static bool32 IsUnionRoomPlayerInvisible(u32 playerIdx, u32 direction)
     return IsObjectEventSpriteInvisible(UR_PLAYER_SPRITE_ID(playerIdx, direction) - UR_SPRITE_START_ID);
 }
 
-static void SpawnGroupMember(u32 playerIdx, u32 direction, u8 graphicsId, struct GFtgtGname * gname)
+static void SpawnGroupMember(u32 playerIdx, u32 direction, u16 graphicsId, struct GFtgtGname * gname)
 {
     s32 x, y;
     s32 id = UR_PLAYER_SPRITE_ID(playerIdx, direction);
